Avoid header from over-shrinking with min-height
Problem: When viewport's height shortened, the header shortened along
with it and even got shorter than the main col. It made the main col
overflow outside the body. It also led to inconsistent height between the
two main columns: header and main.

Solution: Set body's min-height to 900px, so it won't infinitely shrink
and shorter than the main col.